requirements.txt
asttokens==2.2.1
attrs==22.2.0
aws-cdk-lib==2.63.0
aws-cdk.asset-awscli-v1==2.2.56
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.45
backcall==0.2.0
boto3==1.26.63
botocore==1.29.63
cattrs==22.2.0
colorama==0.4.6
comm==0.1.2
constructs==10.1.239
debugpy==1.6.6
decorator==5.1.1
exceptiongroup==1.1.0
executing==1.2.0
greenlet==2.0.2
importlib-metadata==6.0.0
ipykernel==6.21.1
ipython==8.9.0
jedi==0.18.2
jmespath==1.0.1
jsii==1.74.0
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
publication==0.0.3
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.1
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.12.0
